[Feat] #12 : create a Comment component
Create a comment component that will contain the comment form inside it.
The Comment component will handle the states of the comments.
It will also contains the list of comments and pass it to the comment list (tobedeveloped).
